requirements.txt
# Pin versions to match local environment and ensure consistency
paddlepaddle==3.2.2
paddleocr==3.3.2
paddlex==3.3.12
fastapi
uvicorn[standard]
python-multipart
pillow
python-dotenv
opencv-python-headless
